Add model Run
This update adds a Run model to the Django app 'api/v1'. An object of
class Run corresponds to one record in the Runs database table.

The reference to the user id is still missing because I had some
problems to figure out how to add a foreign key field from a module that
is defined in a different app (Users).